+++ src/shared/domain/http/http-error.model.ts
@@ -5,24 +5,24 @@ export class HttpError {
   private static readonly DEFAULT_MESSAGE = 'Something unexpected happened, our technical staff have been automatically notified';
   private static readonly DEFAULT_STATUS = 500;
 
-  timestamp: string;
-  status: number;
-  error: string;
-  exception: string;
-  message: string;
-  path: string;
-  details?: any;
-  callbackErrors?: any;
-  callbackWarnings?: any;
+  public timestamp: string;
+  public status: number;
+  public error: string;
+  public exception: string;
+  public message: string;
+  public path: string;
+  public details?: any;
+  public callbackErrors?: any;
+  public callbackWarnings?: any;
 
-  static from(response: HttpErrorResponse): HttpError {
+  public static from(response: HttpErrorResponse): HttpError {
     const error = new HttpError();
 
     // Check that the HttpErrorResponse contains an "error" object before mapping the error properties
     if (!!(response && response.error)) {
-      Object
-        .keys(error)
-        .forEach(key => error[key] = response.error.hasOwnProperty(key) && response.error[key] ? response.error[key] : error[key]);
+      Object.keys(error).forEach(key => {
+        error[key] = response.error[key] ? response.error[key] : (response[key] || error[key]);
+      });
     }
 
     return error;