+++ src/shared/services/activity/activity.service.ts
@@ -1,10 +1,11 @@
-import { Injectable } from '@angular/core';
-import { Activity } from '../../domain/activity';
-import { Observable } from 'rxjs';
-import { AbstractAppConfig } from '../../../app.config';
-import { HttpService, OptionsType } from '../../services/http';
 import { HttpHeaders } from '@angular/common/http';
-import { SessionStorageService } from '../session/session-storage.service';
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+
+import { AbstractAppConfig } from '../../../app.config';
+import { Activity } from '../../domain/activity';
+import { HttpService, OptionsType } from '../http';
+import { SessionStorageService } from '../session';
 
 // @dynamic
 @Injectable()
@@ -13,11 +14,17 @@ export class ActivityService {
   static get ACTIVITY_VIEW() { return 'view'; }
   static get ACTIVITY_EDIT() { return 'edit'; }
 
-  private userAuthorised;
+  private userAuthorised: boolean = undefined;
 
-  constructor(private readonly http: HttpService,
-              private readonly appConfig: AbstractAppConfig,
-              private readonly sessionStorageService: SessionStorageService) {}
+  public get isEnabled(): boolean {
+    return this.activityUrl() && this.userAuthorised;
+  }
+
+  constructor(
+    private readonly http: HttpService,
+    private readonly appConfig: AbstractAppConfig,
+    private readonly sessionStorageService: SessionStorageService
+  ) {}
 
   public getOptions(): OptionsType {
     const userDetails = JSON.parse(this.sessionStorageService.getItem('userDetails'));
@@ -30,7 +37,7 @@ export class ActivityService {
     return options;
   }
 
-  getActivities(...caseId: string[]): Observable<Activity[]> {
+  public getActivities(...caseId: string[]): Observable<Activity[]> {
     const options = this.getOptions();
     const url = this.activityUrl() + `/cases/${caseId.join(',')}/activity`;
     return this.http
@@ -38,25 +45,25 @@ export class ActivityService {
       .map(response => response);
   }
 
-  postActivity(caseId: string, activityType: String): Observable<Activity[]> {
+  public postActivity(caseId: string, activity: string): Observable<Activity[]> {
     const options = this.getOptions();
     const url = this.activityUrl() + `/cases/${caseId}/activity`;
-    let body = { activity: activityType};
+    let body = { activity };
     return this.http
       .post(url, body, options, false)
       .map(response => response);
   }
 
-  verifyUserIsAuthorized(): void {
+  public verifyUserIsAuthorized(): void {
     if (this.activityUrl() && this.userAuthorised === undefined) {
       this.getActivities(ActivityService.DUMMY_CASE_REFERENCE).subscribe(
-        data => this.userAuthorised = true,
+        () => this.userAuthorised = true,
         error => {
-            if (error.status === 403) {
-              this.userAuthorised = false;
-            } else {
-              this.userAuthorised = true
-            }
+          if ([401, 403].indexOf(error.status) > -1) {
+            this.userAuthorised = false;
+          } else {
+            this.userAuthorised = true
+          }
         }
       );
     }
@@ -64,10 +71,6 @@ export class ActivityService {
 
   private activityUrl(): string {
     return this.appConfig.getActivityUrl();
-  }
-
-  get isEnabled(): boolean {
-    return this.activityUrl() && this.userAuthorised;
   }
 
 }